fix(academic): 修复课表与完成度测试回归
